requirements.txt
beautifulsoup4>=4.12.2
blinker>=1.6.2
certifi>=2023.5.7
charset-normalizer>=3.1.0
cryptography>=40.0.2
click>=8.1.3
dashscope>=1.16.0
Flask>=2.3.2
Flask-SQLAlchemy>=3.0.3
greenlet>=2.0.2
idna>=3.4
itsdangerous>=2.1.2
Jinja2>=3.1.2
MarkupSafe>=2.1.2
numpy>=1.24.3
PyMySQL>=1.0.3
PyYAML>=6.0
redis>=5.0.3
requests>=2.30.0
SQLAlchemy>=2.0.14
typing_extensions>=4.5.0
urllib3>=2.0.2
Werkzeug>=2.3.4

# 部署在Windows下，直接安装opencv-python即可
# opencv-python>=4.7.0.72
# 部署在Linux下，需要使用opencv-python-headless
# opencv-python-headless>=4.7.0.72

# 部署在Linux下时，需要额外安装uWSGI作为WSGI服务器
# uWSGI>=2.0
# 部署在Windows下时，通常使用Tornado作为WSGI服务器
# tornado>=6.1